An admin can not set a user password
Instead of allowing an admin to set the password for a user, the
admin now has a button that they can use to trigger a password
reset for a user, which also sends a notification email to the
user with a reset link.

Additionally, this change introduces a new welcome email that
will be sent when an admin creates a new user and also includes
a password reset link for the new user.

Closes #110