Only update a user password before validations when nil
If we unconditionally run the check, the password will always
be randomly updated when the user saves, as `password` will be
nil _except_ for the case where a user is updating their password.

This change requires that encrypted_password be nil as well before
updating the empty password.

There is a minor change expanding the user of `authenticate_user!`
to reduce potential exceptions being thrown via users poking around
where they should not.

Closes-Bug: #122