Update user experience with admin Reset Password
This change updates the reset password for an admin to include
a confirmation dialog before submitting the reset request,
and additionally adds a reset password button to the User
show page.

There is a bit of additional tidying performed in the
User#force_reset_password! method to ensure that we return
True/False, rather than nil/False, making error propegation
to the admin better.

Closes #141