Improve the admin/index pages to be more functional
This change begins by moving the per-row actions to the left
side of the row, ensuring that the actions are always visible.
Also included is an ability to customize the show vs index page
attributes differently, which lets us include Expires and
Last Activity in the Show page for Users, while discarding
them from the index page.

It then changes the words "Show", "Edit", "destroy" into icons,
and then modifies the base admin controller to allow for disabling
model deletion, and disables deletion for Users.

Closes #150