Resolve conflict between setting and devise dynamic methods
With Setting and Devise sharing the method named `reset_password_within`,
attempts to validate a User's reset token validity would fail when
attempting to call `_value_of` on a User (finally failing when calling
it on an ApplicationRecord (abstract)). By renaming the Setting to
`devise_reset_password_within`, this conflict is avoided and Devise
can successfully check the reset token validity.

Closes #175