Updates to resolve deprecations and warnings (#190)
There were several deprecation notices in the test output
so this change is an update to remove those. Additionally,
there was a warning about possible false negatives in the
controller testing as a result of being too specific in
what exceptions are not expected to be raised so that was
resolved as well.