Fix up Setting updating with optional update params
Previously, several of the options "default" to falsey/nil
when the parameter is not provided for update. This is the
result of a simplistic edit form that did not handle the
falsey/nil value in the actual HTML form, instead defaulting
to falsey/nil when the value was not supplied. This causes
errors after the Settings update pages were split into sections.

Closes #211