It supports bearer tokens for the basic auth handler
Access to management and use of access tokens is restricted to
users of groups that have this feature enabled. When a valid
access token is provided, a 200 is returned from the Nginx
auth request handler, even without an existing user session.
This can be used to allow enabled users to communicate with
APIs behind the Nginx auth request.

Closes #183