Add support for Webauthn authentication (#480)
This change does not include the ability to migrate from a
U2F token to a webauthn token, but registering and using a
new webauthn token should work. Keeping open the option to
enable migration of U2F tokens in the future, this change
also doesn't cleanup the database tables or dependencies
for U2F just yet.